+++ Engine/ScriptCore/src/Components/GameObjectAttached.cs
@@ -9,10 +9,10 @@ namespace Stulu {
 	//default components
 	public class GameObjectAttached {
 		public GameObject gameObject { get; internal set; }
-		internal void initilize(uint goId) {
+		internal void initilize(ulong goId) {
 			this.gameObject = new GameObject(goId);
 		}
-		public uint id { get => gameObject.ID; }
+		public ulong id { get => gameObject.ID; }
 
 		public TransformComponent transform => gameObject.transform;
 		public string name { get => this.gameObject.Name; set => this.gameObject.Name = value; }
